Rewrite tests; separate numba functions for coverage testing
